Split epoch.rs into epoch module and across several files
