+++ src/epoch/mod.rs
@@ -13,21 +13,15 @@ use crate::efmt::format::Format;
 use crate::leap_seconds::{LatestLeapSeconds, LeapSecondProvider};
 use crate::Weekday;
 use crate::{
-    Errors, MonthName, TimeScale, TimeUnits, BDT_REF_EPOCH, DAYS_PER_YEAR_NLD, ET_EPOCH_S,
-    GPST_REF_EPOCH, GST_REF_EPOCH, HIFITIME_REF_YEAR, J1900_OFFSET, J2000_TO_J1900_DURATION,
-    MJD_OFFSET, NANOSECONDS_PER_DAY, NANOSECONDS_PER_MICROSECOND, NANOSECONDS_PER_MILLISECOND,
-    NANOSECONDS_PER_SECOND_U32, QZSST_REF_EPOCH, UNIX_REF_EPOCH,
+    Errors, MonthName, TimeScale, TimeUnits, BDT_REF_EPOCH, ET_EPOCH_S, GPST_REF_EPOCH,
+    GST_REF_EPOCH, J1900_OFFSET, J2000_TO_J1900_DURATION, MJD_OFFSET, NANOSECONDS_PER_DAY,
+    QZSST_REF_EPOCH, UNIX_REF_EPOCH,
 };
 
 #[cfg(not(kani))]
-use crate::parser::Token;
-#[cfg(not(kani))]
 use crate::ParsingErrors;
 
-use core::cmp::{Eq, Ord, Ordering, PartialEq, PartialOrd};
-use core::fmt;
-use core::hash::{Hash, Hasher};
-use core::ops::{Add, AddAssign, Sub, SubAssign};
+use core::cmp::Eq;
 
 #[cfg(feature = "python")]
 mod python;
@@ -57,6 +51,13 @@ mod kani_verif;
 #[cfg(kani)]
 use kani::assert;
 
+mod gregorian;
+pub use gregorian::is_gregorian_valid;
+
+mod formatting;
+mod ops;
+mod with_funcs;
+
 pub(crate) const TT_OFFSET_MS: i64 = 32_184;
 pub(crate) const ET_OFFSET_US: i64 = 32_184_935;
 
@@ -68,61 +69,6 @@ pub const NAIF_M1: f64 = 1.99096871e-7;
 pub const NAIF_EB: f64 = 1.671e-2;
 /// NAIF leap second kernel data used to calculate the difference between ET and TAI.
 pub const NAIF_K: f64 = 1.657e-3;
-
-/// Years when January had the leap second
-const fn january_years(year: i32) -> bool {
-    matches!(
-        year,
-        1972 | 1973
-            | 1974
-            | 1975
-            | 1976
-            | 1977
-            | 1978
-            | 1979
-            | 1980
-            | 1988
-            | 1990
-            | 1991
-            | 1996
-            | 1999
-            | 2006
-            | 2009
-            | 2017
-    )
-}
-
-/// Years when July had the leap second
-const fn july_years(year: i32) -> bool {
-    matches!(
-        year,
-        1972 | 1981 | 1982 | 1983 | 1985 | 1992 | 1993 | 1994 | 1997 | 2012 | 2015
-    )
-}
-
-/// Returns the usual days in a given month (zero indexed, i.e. January is month zero and December is month 11)
-///
-/// # Warning
-/// This will return 0 days if the month is invalid.
-const fn usual_days_per_month(month: u8) -> u8 {
-    match month + 1 {
-        1 | 3 | 5 | 7 | 8 | 10 | 12 => 31,
-        4 | 6 | 9 | 11 => 30,
-        2 => 28,
-        _ => 0,
-    }
-}
-
-/// Calculates the prefix-sum of days counted up to the month start
-const CUMULATIVE_DAYS_FOR_MONTH: [u16; 12] = {
-    let mut days = [0; 12];
-    let mut month = 1;
-    while month < 12 {
-        days[month] = days[month - 1] + usual_days_per_month(month as u8 - 1) as u16;
-        month += 1;
-    }
-    days
-};
 
 /// Defines a nanosecond-precision Epoch.
 ///
@@ -136,13 +82,6 @@ pub struct Epoch {
     pub duration: Duration,
     /// Time scale used during the initialization of this Epoch.
     pub time_scale: TimeScale,
-}
-
-impl Hash for Epoch {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.duration.hash(state);
-        self.time_scale.hash(state);
-    }
 }
 
 #[cfg(not(kani))]
@@ -169,176 +108,8 @@ impl<'de> Deserialize<'de> for Epoch {
     }
 }
 
-impl Sub for Epoch {
-    type Output = Duration;
-
-    fn sub(self, other: Self) -> Duration {
-        self.duration - other.to_time_scale(self.time_scale).duration
-    }
-}
-
-impl SubAssign<Duration> for Epoch {
-    fn sub_assign(&mut self, duration: Duration) {
-        *self = *self - duration;
-    }
-}
-
-impl Sub<Duration> for Epoch {
-    type Output = Self;
-
-    fn sub(self, duration: Duration) -> Self {
-        Self {
-            duration: self.duration - duration,
-            time_scale: self.time_scale,
-        }
-    }
-}
-
-/// WARNING: For speed, there is a possibility to add seconds directly to an Epoch. These will be added in the time scale the Epoch was initialized in.
-/// Using this is _discouraged_ and should only be used if you have facing bottlenecks with the units.
-impl Add<f64> for Epoch {
-    type Output = Self;
-
-    fn add(self, seconds: f64) -> Self {
-        Self {
-            duration: self.duration + seconds * Unit::Second,
-            time_scale: self.time_scale,
-        }
-    }
-}
-
-impl Add<Duration> for Epoch {
-    type Output = Self;
-
-    fn add(self, duration: Duration) -> Self {
-        Self {
-            duration: self.duration + duration,
-            time_scale: self.time_scale,
-        }
-    }
-}
-
-impl AddAssign<Unit> for Epoch {
-    #[allow(clippy::identity_op)]
-    fn add_assign(&mut self, unit: Unit) {
-        *self = *self + unit * 1;
-    }
-}
-
-impl SubAssign<Unit> for Epoch {
-    #[allow(clippy::identity_op)]
-    fn sub_assign(&mut self, unit: Unit) {
-        *self = *self - unit * 1;
-    }
-}
-
-impl Sub<Unit> for Epoch {
-    type Output = Self;
-
-    #[allow(clippy::identity_op)]
-    fn sub(self, unit: Unit) -> Self {
-        Self {
-            duration: self.duration - unit * 1,
-            time_scale: self.time_scale,
-        }
-    }
-}
-
-impl Add<Unit> for Epoch {
-    type Output = Self;
-
-    #[allow(clippy::identity_op)]
-    fn add(self, unit: Unit) -> Self {
-        Self {
-            duration: self.duration + unit * 1,
-            time_scale: self.time_scale,
-        }
-    }
-}
-
-impl AddAssign<Duration> for Epoch {
-    fn add_assign(&mut self, duration: Duration) {
-        *self = *self + duration;
-    }
-}
-
-/// Equality only checks the duration since J1900 match in TAI, because this is how all of the epochs are referenced.
-impl PartialEq for Epoch {
-    fn eq(&self, other: &Self) -> bool {
-        if self.time_scale == other.time_scale {
-            self.duration == other.duration
-        } else {
-            // If one of the two time scales does not include leap seconds,
-            // we always convert the time scale with leap seconds into the
-            // time scale that does NOT have leap seconds.
-            if self.time_scale.uses_leap_seconds() != other.time_scale.uses_leap_seconds() {
-                if self.time_scale.uses_leap_seconds() {
-                    self.to_time_scale(other.time_scale).duration == other.duration
-                } else {
-                    self.duration == other.to_time_scale(self.time_scale).duration
-                }
-            } else {
-                // Otherwise it does not matter
-                self.duration == other.to_time_scale(self.time_scale).duration
-            }
-        }
-    }
-}
-
-impl PartialOrd for Epoch {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(
-            self.duration
-                .cmp(&other.to_time_scale(self.time_scale).duration),
-        )
-    }
-}
-
-impl Ord for Epoch {
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.duration
-            .cmp(&other.to_time_scale(self.time_scale).duration)
-    }
-}
-
 // Defines the methods that should be classmethods in Python, but must be redefined as per https://github.com/PyO3/pyo3/issues/1003#issuecomment-844433346
 impl Epoch {
-    /// Get the accumulated number of leap seconds up to this Epoch from the provided LeapSecondProvider.
-    /// Returns None if the epoch is before 1960, year at which UTC was defined.
-    ///
-    /// # Why does this function return an `Option` when the other returns a value
-    /// This is to match the `iauDat` function of SOFA (src/dat.c). That function will return a warning and give up if the start date is before 1960.
-    pub fn leap_seconds_with<L: LeapSecondProvider>(
-        &self,
-        iers_only: bool,
-        provider: L,
-    ) -> Option<f64> {
-        for leap_second in provider.rev() {
-            if self.to_tai_duration().to_seconds() >= leap_second.timestamp_tai_s
-                && (!iers_only || leap_second.announced_by_iers)
-            {
-                return Some(leap_second.delta_at);
-            }
-        }
-        None
-    }
-
-    /// Creates an epoch from given duration expressed in given timescale, i.e. since the given time scale's reference epoch.
-    ///
-    /// For example, if the duration is 1 day and the time scale is Ephemeris Time, then this will create an epoch of 2000-01-02 at midnight ET. If the duration is 1 day and the time scale is TAI, this will create an epoch of 1900-01-02 at noon, because the TAI reference epoch in Hifitime is chosen to be the J1900 epoch.
-    /// In case of ET, TDB Timescales, a duration since J2000 is expected.
-    #[must_use]
-    pub const fn from_duration(duration: Duration, ts: TimeScale) -> Self {
-        Self {
-            duration,
-            time_scale: ts,
-        }
-    }
-
-    pub fn to_duration_since_j1900(&self) -> Duration {
-        self.to_time_scale(TimeScale::TAI).duration
-    }
-
     #[must_use]
     /// Converts self to another time scale
     ///
@@ -470,6 +241,42 @@ impl Epoch {
             duration,
             time_scale: TimeScale::TAI,
         }
+    }
+
+    /// Get the accumulated number of leap seconds up to this Epoch from the provided LeapSecondProvider.
+    /// Returns None if the epoch is before 1960, year at which UTC was defined.
+    ///
+    /// # Why does this function return an `Option` when the other returns a value
+    /// This is to match the `iauDat` function of SOFA (src/dat.c). That function will return a warning and give up if the start date is before 1960.
+    pub fn leap_seconds_with<L: LeapSecondProvider>(
+        &self,
+        iers_only: bool,
+        provider: L,
+    ) -> Option<f64> {
+        for leap_second in provider.rev() {
+            if self.to_tai_duration().to_seconds() >= leap_second.timestamp_tai_s
+                && (!iers_only || leap_second.announced_by_iers)
+            {
+                return Some(leap_second.delta_at);
+            }
+        }
+        None
+    }
+
+    /// Creates an epoch from given duration expressed in given timescale, i.e. since the given time scale's reference epoch.
+    ///
+    /// For example, if the duration is 1 day and the time scale is Ephemeris Time, then this will create an epoch of 2000-01-02 at midnight ET. If the duration is 1 day and the time scale is TAI, this will create an epoch of 1900-01-02 at noon, because the TAI reference epoch in Hifitime is chosen to be the J1900 epoch.
+    /// In case of ET, TDB Timescales, a duration since J2000 is expected.
+    #[must_use]
+    pub const fn from_duration(duration: Duration, ts: TimeScale) -> Self {
+        Self {
+            duration,
+            time_scale: ts,
+        }
+    }
+
+    pub fn to_duration_since_j1900(&self) -> Duration {
+        self.to_time_scale(TimeScale::TAI).duration
     }
 
     #[must_use]
@@ -823,404 +630,6 @@ impl Epoch {
         Self::from_utc_duration(UNIX_REF_EPOCH.to_utc_duration() + millisecond * Unit::Millisecond)
     }
 
-    /// Attempts to build an Epoch from the provided Gregorian date and time in TAI.
-    pub fn maybe_from_gregorian_tai(
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-        nanos: u32,
-    ) -> Result<Self, Errors> {
-        Self::maybe_from_gregorian(
-            year,
-            month,
-            day,
-            hour,
-            minute,
-            second,
-            nanos,
-            TimeScale::TAI,
-        )
-    }
-
-    /// Attempts to build an Epoch from the provided Gregorian date and time in the provided time scale.
-    /// NOTE: If the time scale is TDB, this function assumes that the SPICE format is used
-    #[allow(clippy::too_many_arguments)]
-    pub fn maybe_from_gregorian(
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-        nanos: u32,
-        time_scale: TimeScale,
-    ) -> Result<Self, Errors> {
-        if !is_gregorian_valid(year, month, day, hour, minute, second, nanos) {
-            return Err(Errors::Carry);
-        }
-
-        let mut duration_wrt_ref = match year.checked_sub(HIFITIME_REF_YEAR) {
-            None => return Err(Errors::Overflow),
-            Some(years_since_ref) => match years_since_ref.checked_mul(365) {
-                None => return Err(Errors::Overflow),
-                Some(days) => Unit::Day * i64::from(days),
-            },
-        } - time_scale.gregorian_epoch_offset();
-
-        // Now add the leap days for all the years prior to the current year
-        if year >= HIFITIME_REF_YEAR {
-            // Add days
-            for year in HIFITIME_REF_YEAR..year {
-                if is_leap_year(year) {
-                    duration_wrt_ref += Unit::Day;
-                }
-            }
-            // Remove ref hours from duration to correct for the time scale not starting at midnight
-            // duration_wrt_ref -= Unit::Hour * time_scale.ref_hour() as i64;
-        } else {
-            // Remove days
-            for year in year..HIFITIME_REF_YEAR {
-                if is_leap_year(year) {
-                    duration_wrt_ref -= Unit::Day;
-                }
-            }
-            // Add ref hours
-            // duration_wrt_ref += Unit::Hour * time_scale.ref_hour() as i64;
-        }
-
-        // Add the seconds for the months prior to the current month
-        duration_wrt_ref += Unit::Day * i64::from(CUMULATIVE_DAYS_FOR_MONTH[(month - 1) as usize]);
-
-        if is_leap_year(year) && month > 2 {
-            // NOTE: If on 29th of February, then the day is not finished yet, and therefore
-            // the extra seconds are added below as per a normal day.
-            duration_wrt_ref += Unit::Day;
-        }
-        duration_wrt_ref += Unit::Day * i64::from(day - 1)
-            + Unit::Hour * i64::from(hour)
-            + Unit::Minute * i64::from(minute)
-            + Unit::Second * i64::from(second)
-            + Unit::Nanosecond * i64::from(nanos);
-
-        if second == 60 {
-            // Herein lies the whole ambiguity of leap seconds. Two different UTC dates exist at the
-            // same number of second after J1900.0.
-            duration_wrt_ref -= Unit::Second;
-        }
-
-        Ok(Self {
-            duration: duration_wrt_ref,
-            time_scale,
-        })
-    }
-
-    #[must_use]
-    /// Builds an Epoch from the provided Gregorian date and time in TAI. If invalid date is provided, this function will panic.
-    /// Use maybe_from_gregorian_tai if unsure.
-    pub fn from_gregorian_tai(
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-        nanos: u32,
-    ) -> Self {
-        Self::maybe_from_gregorian_tai(year, month, day, hour, minute, second, nanos)
-            .expect("invalid Gregorian date")
-    }
-
-    #[must_use]
-    /// Initialize from the Gregorian date at midnight in TAI.
-    pub fn from_gregorian_tai_at_midnight(year: i32, month: u8, day: u8) -> Self {
-        Self::maybe_from_gregorian_tai(year, month, day, 0, 0, 0, 0)
-            .expect("invalid Gregorian date")
-    }
-
-    #[must_use]
-    /// Initialize from the Gregorian date at noon in TAI
-    pub fn from_gregorian_tai_at_noon(year: i32, month: u8, day: u8) -> Self {
-        Self::maybe_from_gregorian_tai(year, month, day, 12, 0, 0, 0)
-            .expect("invalid Gregorian date")
-    }
-
-    #[must_use]
-    /// Initialize from the Gregorian date and time (without the nanoseconds) in TAI
-    pub fn from_gregorian_tai_hms(
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-    ) -> Self {
-        Self::maybe_from_gregorian_tai(year, month, day, hour, minute, second, 0)
-            .expect("invalid Gregorian date")
-    }
-
-    /// Attempts to build an Epoch from the provided Gregorian date and time in UTC.
-    pub fn maybe_from_gregorian_utc(
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-        nanos: u32,
-    ) -> Result<Self, Errors> {
-        Self::maybe_from_gregorian(
-            year,
-            month,
-            day,
-            hour,
-            minute,
-            second,
-            nanos,
-            TimeScale::UTC,
-        )
-    }
-
-    #[must_use]
-    /// Builds an Epoch from the provided Gregorian date and time in UTC. If invalid date is provided, this function will panic.
-    /// Use maybe_from_gregorian_utc if unsure.
-    pub fn from_gregorian_utc(
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-        nanos: u32,
-    ) -> Self {
-        Self::maybe_from_gregorian_utc(year, month, day, hour, minute, second, nanos)
-            .expect("invalid Gregorian date")
-    }
-
-    #[must_use]
-    /// Initialize from Gregorian date in UTC at midnight
-    pub fn from_gregorian_utc_at_midnight(year: i32, month: u8, day: u8) -> Self {
-        Self::maybe_from_gregorian_utc(year, month, day, 0, 0, 0, 0)
-            .expect("invalid Gregorian date")
-    }
-
-    #[must_use]
-    /// Initialize from Gregorian date in UTC at noon
-    pub fn from_gregorian_utc_at_noon(year: i32, month: u8, day: u8) -> Self {
-        Self::maybe_from_gregorian_utc(year, month, day, 12, 0, 0, 0)
-            .expect("invalid Gregorian date")
-    }
-
-    #[must_use]
-    /// Initialize from the Gregorian date and time (without the nanoseconds) in UTC
-    pub fn from_gregorian_utc_hms(
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-    ) -> Self {
-        Self::maybe_from_gregorian_utc(year, month, day, hour, minute, second, 0)
-            .expect("invalid Gregorian date")
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    #[must_use]
-    /// Builds an Epoch from the provided Gregorian date and time in the provided time scale. If invalid date is provided, this function will panic.
-    /// Use maybe_from_gregorian if unsure.
-    pub fn from_gregorian(
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-        nanos: u32,
-        time_scale: TimeScale,
-    ) -> Self {
-        Self::maybe_from_gregorian(year, month, day, hour, minute, second, nanos, time_scale)
-            .expect("invalid Gregorian date")
-    }
-
-    #[must_use]
-    /// Initialize from Gregorian date in UTC at midnight
-    pub fn from_gregorian_at_midnight(
-        year: i32,
-        month: u8,
-        day: u8,
-        time_scale: TimeScale,
-    ) -> Self {
-        Self::maybe_from_gregorian(year, month, day, 0, 0, 0, 0, time_scale)
-            .expect("invalid Gregorian date")
-    }
-
-    #[must_use]
-    /// Initialize from Gregorian date in UTC at noon
-    pub fn from_gregorian_at_noon(year: i32, month: u8, day: u8, time_scale: TimeScale) -> Self {
-        Self::maybe_from_gregorian(year, month, day, 12, 0, 0, 0, time_scale)
-            .expect("invalid Gregorian date")
-    }
-
-    #[must_use]
-    /// Initialize from the Gregorian date and time (without the nanoseconds) in UTC
-    pub fn from_gregorian_hms(
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-        time_scale: TimeScale,
-    ) -> Self {
-        Self::maybe_from_gregorian(year, month, day, hour, minute, second, 0, time_scale)
-            .expect("invalid Gregorian date")
-    }
-
-    /// Converts a Gregorian date time in ISO8601 or RFC3339 format into an Epoch, accounting for the time zone designator and the time scale.
-    ///
-    /// # Definition
-    /// 1. Time Zone Designator: this is either a `Z` (lower or upper case) to specify UTC, or an offset in hours and minutes off of UTC, such as `+01:00` for UTC plus one hour and zero minutes.
-    /// 2. Time system (or time "scale"): UTC, TT, TAI, TDB, ET, etc.
-    ///
-    /// Converts an ISO8601 or RFC3339 datetime representation to an Epoch.
-    /// If no time scale is specified, then UTC is assumed.
-    /// A time scale may be specified _in addition_ to the format unless
-    /// The `T` which separates the date from the time can be replaced with a single whitespace character (`\W`).
-    /// The offset is also optional, cf. the examples below.
-    ///
-    /// # Example
-    /// ```
-    /// use hifitime::Epoch;
-    /// let dt = Epoch::from_gregorian_utc(2017, 1, 14, 0, 31, 55, 0);
-    /// assert_eq!(
-    ///     dt,
-    ///     Epoch::from_gregorian_str("2017-01-14T00:31:55 UTC").unwrap()
-    /// );
-    /// assert_eq!(
-    ///     dt,
-    ///     Epoch::from_gregorian_str("2017-01-14T00:31:55.0000 UTC").unwrap()
-    /// );
-    /// assert_eq!(
-    ///     dt,
-    ///     Epoch::from_gregorian_str("2017-01-14T00:31:55").unwrap()
-    /// );
-    /// assert_eq!(
-    ///     dt,
-    ///     Epoch::from_gregorian_str("2017-01-14 00:31:55").unwrap()
-    /// );
-    /// // Regression test for #90
-    /// assert_eq!(
-    ///     Epoch::from_gregorian_utc(2017, 1, 14, 0, 31, 55, 811000000),
-    ///     Epoch::from_gregorian_str("2017-01-14 00:31:55.811 UTC").unwrap()
-    /// );
-    /// assert_eq!(
-    ///     Epoch::from_gregorian_utc(2017, 1, 14, 0, 31, 55, 811200000),
-    ///     Epoch::from_gregorian_str("2017-01-14 00:31:55.8112 UTC").unwrap()
-    /// );
-    /// // Example from https://www.w3.org/TR/NOTE-datetime
-    /// assert_eq!(
-    ///     Epoch::from_gregorian_utc_hms(1994, 11, 5, 13, 15, 30),
-    ///     Epoch::from_gregorian_str("1994-11-05T13:15:30Z").unwrap()
-    /// );
-    /// assert_eq!(
-    ///     Epoch::from_gregorian_utc_hms(1994, 11, 5, 13, 15, 30),
-    ///     Epoch::from_gregorian_str("1994-11-05T08:15:30-05:00").unwrap()
-    /// );
-    /// ```
-    #[cfg(not(kani))]
-    pub fn from_gregorian_str(s_in: &str) -> Result<Self, Errors> {
-        // All of the integers in a date: year, month, day, hour, minute, second, subsecond, offset hours, offset minutes
-        let mut decomposed = [0_i32; 9];
-        // The parsed time scale, defaults to UTC
-        let mut ts = TimeScale::UTC;
-        // The offset sign, defaults to positive.
-        let mut offset_sign = 1;
-
-        // Previous index of interest in the string
-        let mut prev_idx = 0;
-        let mut cur_token = Token::Year;
-
-        let s = s_in.trim();
-
-        for (idx, char) in s.chars().enumerate() {
-            if !char.is_numeric() || idx == s.len() - 1 {
-                if cur_token == Token::Timescale {
-                    // Then we match the timescale directly.
-                    if idx != s.len() - 1 {
-                        // We have some remaining characters, so let's parse those in the only formats we know.
-                        ts = TimeScale::from_str(s[idx..].trim())?;
-                    }
-                    break;
-                }
-                let prev_token = cur_token;
-
-                let pos = cur_token.gregorian_position().unwrap();
-
-                let end_idx = if idx != s.len() - 1 || !char.is_numeric() {
-                    // Only advance the token if we aren't at the end of the string
-                    cur_token.advance_with(char)?;
-                    idx
-                } else {
-                    idx + 1
-                };
-
-                if prev_idx > end_idx {
-                    return Err(Errors::ParseError(ParsingErrors::ISO8601));
-                }
-
-                match lexical_core::parse(s[prev_idx..end_idx].as_bytes()) {
-                    Ok(val) => {
-                        // Check that this valid is OK for the token we're reading it as.
-                        prev_token.value_ok(val)?;
-                        // If these are the subseconds, we must convert them to nanoseconds
-                        if prev_token == Token::Subsecond {
-                            if end_idx - prev_idx != 9 {
-                                decomposed[pos] =
-                                    val * 10_i32.pow((9 - (end_idx - prev_idx)) as u32);
-                            } else {
-                                decomposed[pos] = val;
-                            }
-                        } else {
-                            decomposed[pos] = val
-                        }
-                    }
-                    Err(_) => return Err(Errors::ParseError(ParsingErrors::ISO8601)),
-                }
-                prev_idx = idx + 1;
-                // If we are about to parse an hours offset, we need to set the sign now.
-                if cur_token == Token::OffsetHours {
-                    if &s[idx..idx + 1] == "-" {
-                        offset_sign = -1;
-                    }
-                    prev_idx += 1;
-                }
-            }
-        }
-
-        let tz = if offset_sign > 0 {
-            // We oppose the sign in the string to undo the offset
-            -(i64::from(decomposed[7]) * Unit::Hour + i64::from(decomposed[8]) * Unit::Minute)
-        } else {
-            i64::from(decomposed[7]) * Unit::Hour + i64::from(decomposed[8]) * Unit::Minute
-        };
-
-        let epoch = Self::maybe_from_gregorian(
-            decomposed[0],
-            decomposed[1].try_into().unwrap(),
-            decomposed[2].try_into().unwrap(),
-            decomposed[3].try_into().unwrap(),
-            decomposed[4].try_into().unwrap(),
-            decomposed[5].try_into().unwrap(),
-            decomposed[6].try_into().unwrap(),
-            ts,
-        )?;
-
-        Ok(epoch + tz)
-    }
-
     /// Initializes an Epoch from the provided Format.
     pub fn from_str_with_format(s_in: &str, format: Format) -> Result<Self, Errors> {
         format.parse(s_in)
@@ -1247,135 +656,6 @@ impl Epoch {
         let g = TAU / 360.0 * 357.528 + 1.990_910_018_065_731e-7 * seconds;
         // Return gamma
         1.658e-3 * (g + 1.67e-2 * g.sin()).sin()
-    }
-
-    pub(crate) fn compute_gregorian(
-        duration: Duration,
-        ts: TimeScale,
-    ) -> (i32, u8, u8, u8, u8, u8, u32) {
-        let (sign, days, mut hours, minutes, seconds, milliseconds, microseconds, nanos) =
-            (duration + ts.gregorian_epoch_offset()).decompose();
-
-        let days_f64 = if sign < 0 {
-            -(days as f64)
-        } else {
-            days as f64
-        };
-
-        let (mut year, mut days_in_year) = div_rem_f64(days_f64, DAYS_PER_YEAR_NLD);
-        year += HIFITIME_REF_YEAR;
-
-        // Base calculation was on 365 days, so we need to remove one day in seconds per leap year
-        // between 1900 and `year`
-        if year >= HIFITIME_REF_YEAR {
-            for year in HIFITIME_REF_YEAR..year {
-                if is_leap_year(year) {
-                    days_in_year -= 1.0;
-                }
-            }
-        } else {
-            for year in year..HIFITIME_REF_YEAR {
-                if is_leap_year(year) {
-                    days_in_year += 1.0;
-                }
-            }
-        }
-
-        // Get the month from the exact number of seconds between the start of the year and now
-        let mut month = 1;
-        let mut day;
-
-        let mut days_so_far = 0.0;
-        loop {
-            let mut days_next_month = usual_days_per_month(month - 1) as f64;
-            if month == 2 && is_leap_year(year) {
-                days_next_month += 1.0;
-            }
-
-            if days_so_far + days_next_month > days_in_year || month == 12 {
-                // We've found the month and can calculate the days
-                day = if sign >= 0 {
-                    days_in_year - days_so_far + 1.0
-                } else {
-                    days_in_year - days_so_far
-                };
-                break;
-            }
-
-            // Otherwise, count up the number of days this year so far and keep track of the month.
-            days_so_far += days_next_month;
-            month += 1;
-        }
-
-        if hours >= 24 {
-            hours -= 24;
-            if year >= HIFITIME_REF_YEAR {
-                day += 1.0;
-            } else {
-                day -= 1.0;
-            }
-        }
-
-        if day <= 0.0 || days_in_year < 0.0 {
-            // We've overflowed backward
-            month = 12;
-            year -= 1;
-            // NOTE: Leap year is already accounted for in the TAI duration when counting backward.
-            day = if days_in_year < 0.0 {
-                days_in_year + usual_days_per_month(11) as f64 + 1.0
-            } else {
-                usual_days_per_month(11) as f64
-            };
-        }
-
-        if sign < 0 {
-            let time = Duration::compose(
-                sign,
-                0,
-                hours,
-                minutes,
-                seconds,
-                milliseconds,
-                microseconds,
-                nanos,
-            );
-
-            // Last check on the validity of the Gregorian date
-
-            if time == Duration::ZERO || month == 12 && day == 32.0 {
-                // We've underflowed since we're before 1900.
-                year += 1;
-                month = 1;
-                day = 1.0;
-            }
-
-            let (_, _, hours, minutes, seconds, milliseconds, microseconds, nanos) =
-                (24 * Unit::Hour + time).decompose();
-
-            (
-                year,
-                month,
-                day as u8,
-                hours as u8,
-                minutes as u8,
-                seconds as u8,
-                (nanos
-                    + microseconds * NANOSECONDS_PER_MICROSECOND
-                    + milliseconds * NANOSECONDS_PER_MILLISECOND) as u32,
-            )
-        } else {
-            (
-                year,
-                month,
-                day as u8,
-                hours as u8,
-                minutes as u8,
-                seconds as u8,
-                (nanos
-                    + microseconds * NANOSECONDS_PER_MICROSECOND
-                    + milliseconds * NANOSECONDS_PER_MILLISECOND) as u32,
-            )
-        }
     }
 
     /// Builds an Epoch from given `week`: elapsed weeks counter into the desired Time scale, and the amount of nanoseconds within that week.
@@ -1463,23 +743,6 @@ impl Epoch {
         }
     }
 
-    /*
-        /// Makes a copy of self and sets the duration and time scale appropriately given the new duration
-        #[must_use]
-        pub fn set(&self, new_duration: Duration) -> Self {
-            match self.time_scale {
-                TimeScale::TAI => Self::from_tai_duration(new_duration),
-                TimeScale::TT => Self::from_tt_duration(new_duration),
-                TimeScale::ET => Self::from_et_duration(new_duration),
-                TimeScale::TDB => Self::from_tdb_duration(new_duration),
-                TimeScale::UTC => Self::from_utc_duration(new_duration),
-                // GPST and QZSST share the same properties
-                TimeScale::GPST | TimeScale::QZSST => Self::from_gpst_duration(new_duration),
-                TimeScale::GST => Self::from_gst_duration(new_duration),
-                TimeScale::BDT => Self::from_bdt_duration(new_duration),
-            }
-        }
-    */
     #[must_use]
     /// Returns the number of TAI seconds since J1900
     pub fn to_tai_seconds(&self) -> f64 {
@@ -1913,232 +1176,6 @@ impl Epoch {
     }
 
     #[must_use]
-    /// Converts the Epoch to the Gregorian UTC equivalent as (year, month, day, hour, minute, second).
-    /// WARNING: Nanoseconds are lost in this conversion!
-    ///
-    /// # Example
-    /// ```
-    /// use hifitime::Epoch;
-    ///
-    /// let dt = Epoch::from_tai_parts(1, 537582752000000000);
-    ///
-    /// // With the std feature, you may use FromStr as such
-    /// // let dt_str = "2017-01-14T00:31:55 UTC";
-    /// // let dt = Epoch::from_gregorian_str(dt_str).unwrap()
-    ///
-    /// let (y, m, d, h, min, s, _) = dt.to_gregorian_utc();
-    /// assert_eq!(y, 2017);
-    /// assert_eq!(m, 1);
-    /// assert_eq!(d, 14);
-    /// assert_eq!(h, 0);
-    /// assert_eq!(min, 31);
-    /// assert_eq!(s, 55);
-    /// #[cfg(feature = "std")]
-    /// assert_eq!("2017-01-14T00:31:55 UTC", format!("{dt}"));
-    /// ```
-    pub fn to_gregorian_utc(&self) -> (i32, u8, u8, u8, u8, u8, u32) {
-        let ts = TimeScale::UTC;
-        Self::compute_gregorian(self.to_duration_in_time_scale(ts), ts)
-    }
-
-    #[must_use]
-    /// Converts the Epoch to the Gregorian TAI equivalent as (year, month, day, hour, minute, second).
-    /// WARNING: Nanoseconds are lost in this conversion!
-    ///
-    /// # Example
-    /// ```
-    /// use hifitime::Epoch;
-    /// let dt = Epoch::from_gregorian_tai_at_midnight(1972, 1, 1);
-    /// let (y, m, d, h, min, s, _) = dt.to_gregorian_tai();
-    /// assert_eq!(y, 1972);
-    /// assert_eq!(m, 1);
-    /// assert_eq!(d, 1);
-    /// assert_eq!(h, 0);
-    /// assert_eq!(min, 0);
-    /// assert_eq!(s, 0);
-    /// ```
-    pub fn to_gregorian_tai(&self) -> (i32, u8, u8, u8, u8, u8, u32) {
-        let ts = TimeScale::TAI;
-        Self::compute_gregorian(self.to_duration_in_time_scale(ts), ts)
-    }
-
-    #[must_use]
-    /// Floors this epoch to the closest provided duration
-    ///
-    /// # Example
-    /// ```
-    /// use hifitime::{Epoch, TimeUnits};
-    ///
-    /// let e = Epoch::from_gregorian_tai_hms(2022, 5, 20, 17, 57, 43);
-    /// assert_eq!(
-    ///     e.floor(1.hours()),
-    ///     Epoch::from_gregorian_tai_hms(2022, 5, 20, 17, 0, 0)
-    /// );
-    ///
-    /// let e = Epoch::from_gregorian_tai(2022, 10, 3, 17, 44, 29, 898032665);
-    /// assert_eq!(
-    ///     e.floor(3.minutes()),
-    ///     Epoch::from_gregorian_tai_hms(2022, 10, 3, 17, 42, 0)
-    /// );
-    /// ```
-    pub fn floor(&self, duration: Duration) -> Self {
-        Self::from_duration(self.duration.floor(duration), self.time_scale)
-    }
-
-    #[must_use]
-    /// Ceils this epoch to the closest provided duration in the TAI time scale
-    ///
-    /// # Example
-    /// ```
-    /// use hifitime::{Epoch, TimeUnits};
-    ///
-    /// let e = Epoch::from_gregorian_tai_hms(2022, 5, 20, 17, 57, 43);
-    /// assert_eq!(
-    ///     e.ceil(1.hours()),
-    ///     Epoch::from_gregorian_tai_hms(2022, 5, 20, 18, 0, 0)
-    /// );
-    ///
-    /// // 45 minutes is a multiple of 3 minutes, hence this result
-    /// let e = Epoch::from_gregorian_tai(2022, 10, 3, 17, 44, 29, 898032665);
-    /// assert_eq!(
-    ///     e.ceil(3.minutes()),
-    ///     Epoch::from_gregorian_tai_hms(2022, 10, 3, 17, 45, 0)
-    /// );
-    /// ```
-    pub fn ceil(&self, duration: Duration) -> Self {
-        Self::from_duration(self.duration.ceil(duration), self.time_scale)
-    }
-
-    #[must_use]
-    /// Rounds this epoch to the closest provided duration in TAI
-    ///
-    /// # Example
-    /// ```
-    /// use hifitime::{Epoch, TimeUnits};
-    ///
-    /// let e = Epoch::from_gregorian_tai_hms(2022, 5, 20, 17, 57, 43);
-    /// assert_eq!(
-    ///     e.round(1.hours()),
-    ///     Epoch::from_gregorian_tai_hms(2022, 5, 20, 18, 0, 0)
-    /// );
-    /// ```
-    pub fn round(&self, duration: Duration) -> Self {
-        Self::from_duration(self.duration.round(duration), self.time_scale)
-    }
-
-    #[must_use]
-    /// Converts this epoch into the time of week, represented as a rolling week counter into that time scale
-    /// and the number of nanoseconds elapsed in current week (since closest Sunday midnight).
-    /// This is usually how GNSS receivers describe a timestamp.
-    pub fn to_time_of_week(&self) -> (u32, u64) {
-        let total_nanoseconds = self.duration.total_nanoseconds();
-        let weeks = total_nanoseconds / NANOSECONDS_PER_DAY as i128 / Weekday::DAYS_PER_WEEK_I128;
-        // elapsed nanoseconds in current week:
-        //   remove previously determined nb of weeks
-        //   get residual nanoseconds
-        let nanoseconds =
-            total_nanoseconds - weeks * NANOSECONDS_PER_DAY as i128 * Weekday::DAYS_PER_WEEK_I128;
-        (weeks as u32, nanoseconds as u64)
-    }
-
-    #[must_use]
-    /// Returns the weekday in provided time scale **ASSUMING** that the reference epoch of that time scale is a Monday.
-    /// You _probably_ do not want to use this. You probably either want `weekday()` or `weekday_utc()`.
-    /// Several time scales do _not_ have a reference day that's on a Monday, e.g. BDT.
-    pub fn weekday_in_time_scale(&self, time_scale: TimeScale) -> Weekday {
-        (rem_euclid_f64(
-            self.to_duration_in_time_scale(time_scale)
-                .to_unit(Unit::Day),
-            Weekday::DAYS_PER_WEEK,
-        )
-        .floor() as u8)
-            .into()
-    }
-
-    #[must_use]
-    /// Returns weekday (uses the TAI representation for this calculation).
-    pub fn weekday(&self) -> Weekday {
-        // J1900 was a Monday so we just have to modulo the number of days by the number of days per week.
-        // The function call will be optimized away.
-        self.weekday_in_time_scale(TimeScale::TAI)
-    }
-
-    #[must_use]
-    /// Returns weekday in UTC timescale
-    pub fn weekday_utc(&self) -> Weekday {
-        self.weekday_in_time_scale(TimeScale::UTC)
-    }
-
-    #[must_use]
-    /// Returns the next weekday.
-    ///
-    /// ```
-    /// use hifitime::prelude::*;
-    ///
-    /// let epoch = Epoch::from_gregorian_utc_at_midnight(1988, 1, 2);
-    /// assert_eq!(epoch.weekday_utc(), Weekday::Saturday);
-    /// assert_eq!(epoch.next(Weekday::Sunday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 3));
-    /// assert_eq!(epoch.next(Weekday::Monday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 4));
-    /// assert_eq!(epoch.next(Weekday::Tuesday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 5));
-    /// assert_eq!(epoch.next(Weekday::Wednesday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 6));
-    /// assert_eq!(epoch.next(Weekday::Thursday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 7));
-    /// assert_eq!(epoch.next(Weekday::Friday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 8));
-    /// assert_eq!(epoch.next(Weekday::Saturday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 9));
-    /// ```
-    pub fn next(&self, weekday: Weekday) -> Self {
-        let delta_days = self.weekday() - weekday;
-        if delta_days == Duration::ZERO {
-            *self + 7 * Unit::Day
-        } else {
-            *self + delta_days
-        }
-    }
-
-    #[must_use]
-    pub fn next_weekday_at_midnight(&self, weekday: Weekday) -> Self {
-        self.next(weekday).with_hms_strict(0, 0, 0)
-    }
-
-    #[must_use]
-    pub fn next_weekday_at_noon(&self, weekday: Weekday) -> Self {
-        self.next(weekday).with_hms_strict(12, 0, 0)
-    }
-
-    #[must_use]
-    /// Returns the next weekday.
-    ///
-    /// ```
-    /// use hifitime::prelude::*;
-    ///
-    /// let epoch = Epoch::from_gregorian_utc_at_midnight(1988, 1, 2);
-    /// assert_eq!(epoch.previous(Weekday::Friday), Epoch::from_gregorian_utc_at_midnight(1988, 1, 1));
-    /// assert_eq!(epoch.previous(Weekday::Thursday), Epoch::from_gregorian_utc_at_midnight(1987, 12, 31));
-    /// assert_eq!(epoch.previous(Weekday::Wednesday), Epoch::from_gregorian_utc_at_midnight(1987, 12, 30));
-    /// assert_eq!(epoch.previous(Weekday::Tuesday), Epoch::from_gregorian_utc_at_midnight(1987, 12, 29));
-    /// assert_eq!(epoch.previous(Weekday::Monday), Epoch::from_gregorian_utc_at_midnight(1987, 12, 28));
-    /// assert_eq!(epoch.previous(Weekday::Sunday), Epoch::from_gregorian_utc_at_midnight(1987, 12, 27));
-    /// assert_eq!(epoch.previous(Weekday::Saturday), Epoch::from_gregorian_utc_at_midnight(1987, 12, 26));
-    /// ```
-    pub fn previous(&self, weekday: Weekday) -> Self {
-        let delta_days = weekday - self.weekday();
-        if delta_days == Duration::ZERO {
-            *self - 7 * Unit::Day
-        } else {
-            *self - delta_days
-        }
-    }
-
-    #[must_use]
-    pub fn previous_weekday_at_midnight(&self, weekday: Weekday) -> Self {
-        self.previous(weekday).with_hms_strict(0, 0, 0)
-    }
-
-    #[must_use]
-    pub fn previous_weekday_at_noon(&self, weekday: Weekday) -> Self {
-        self.previous(weekday).with_hms_strict(12, 0, 0)
-    }
-
-    #[must_use]
     /// Returns the duration since the start of the year
     pub fn duration_in_year(&self) -> Duration {
         let start_of_year = Self::from_gregorian(self.year(), 1, 1, 0, 0, 0, 0, self.time_scale);
@@ -2193,166 +1230,9 @@ impl Epoch {
         self.duration.decompose().7
     }
 
-    /// Returns a copy of self where the time is set to the provided hours, minutes, seconds
-    /// Invalid number of hours, minutes, and seconds will overflow into their higher unit.
-    /// Warning: this does _not_ set the subdivisions of second to zero.
-    pub fn with_hms(&self, hours: u64, minutes: u64, seconds: u64) -> Self {
-        let (sign, days, _, _, _, milliseconds, microseconds, nanoseconds) =
-            self.duration.decompose();
-        Self::from_duration(
-            Duration::compose(
-                sign,
-                days,
-                hours,
-                minutes,
-                seconds,
-                milliseconds,
-                microseconds,
-                nanoseconds,
-            ),
-            self.time_scale,
-        )
-    }
-
-    /// Returns a copy of self where the hours, minutes, seconds is set to the time of the provided epoch but the
-    /// sub-second parts are kept from the current epoch.
-    ///
-    /// ```
-    /// use hifitime::prelude::*;
-    ///
-    /// let epoch = Epoch::from_gregorian_utc(2022, 12, 01, 10, 11, 12, 13);
-    /// let other_utc = Epoch::from_gregorian_utc(2024, 12, 01, 20, 21, 22, 23);
-    /// let other = other_utc.to_time_scale(TimeScale::TDB);
-    ///
-    /// assert_eq!(
-    ///     epoch.with_hms_from(other),
-    ///     Epoch::from_gregorian_utc(2022, 12, 01, 20, 21, 22, 13)
-    /// );
-    /// ```
-    pub fn with_hms_from(&self, other: Self) -> Self {
-        let (sign, days, _, _, _, milliseconds, microseconds, nanoseconds) =
-            self.duration.decompose();
-        // Shadow other with the provided other epoch but in the correct time scale.
-        let other = other.to_time_scale(self.time_scale);
-        Self::from_duration(
-            Duration::compose(
-                sign,
-                days,
-                other.hours(),
-                other.minutes(),
-                other.seconds(),
-                milliseconds,
-                microseconds,
-                nanoseconds,
-            ),
-            self.time_scale,
-        )
-    }
-
-    /// Returns a copy of self where all of the time components (hours, minutes, seconds, and sub-seconds) are set to the time of the provided epoch.
-    ///
-    /// ```
-    /// use hifitime::prelude::*;
-    ///
-    /// let epoch = Epoch::from_gregorian_utc(2022, 12, 01, 10, 11, 12, 13);
-    /// let other_utc = Epoch::from_gregorian_utc(2024, 12, 01, 20, 21, 22, 23);
-    /// // If the other Epoch is in another time scale, it does not matter, it will be converted to the correct time scale.
-    /// let other = other_utc.to_time_scale(TimeScale::TDB);
-    ///
-    /// assert_eq!(
-    ///     epoch.with_time_from(other),
-    ///     Epoch::from_gregorian_utc(2022, 12, 01, 20, 21, 22, 23)
-    /// );
-    /// ```
-    pub fn with_time_from(&self, other: Self) -> Self {
-        // Grab days from self
-        let (sign, days, _, _, _, _, _, _) = self.duration.decompose();
-
-        // Grab everything else from other
-        let (_, _, hours, minutes, seconds, milliseconds, microseconds, nanoseconds) =
-            other.to_duration_in_time_scale(self.time_scale).decompose();
-
-        Self::from_duration(
-            Duration::compose(
-                sign,
-                days,
-                hours,
-                minutes,
-                seconds,
-                milliseconds,
-                microseconds,
-                nanoseconds,
-            ),
-            self.time_scale,
-        )
-    }
-
-    /// Returns a copy of self where the time is set to the provided hours, minutes, seconds
-    /// Invalid number of hours, minutes, and seconds will overflow into their higher unit.
-    /// Warning: this will set the subdivisions of seconds to zero.
-    pub fn with_hms_strict(&self, hours: u64, minutes: u64, seconds: u64) -> Self {
-        let (sign, days, _, _, _, _, _, _) = self.duration.decompose();
-        Self::from_duration(
-            Duration::compose(sign, days, hours, minutes, seconds, 0, 0, 0),
-            self.time_scale,
-        )
-    }
-
-    /// Returns a copy of self where the time is set to the time of the other epoch but the subseconds are set to zero.
-    ///
-    /// ```
-    /// use hifitime::prelude::*;
-    ///
-    /// let epoch = Epoch::from_gregorian_utc(2022, 12, 01, 10, 11, 12, 13);
-    /// let other_utc = Epoch::from_gregorian_utc(2024, 12, 01, 20, 21, 22, 23);
-    /// let other = other_utc.to_time_scale(TimeScale::TDB);
-    ///
-    /// assert_eq!(
-    ///     epoch.with_hms_strict_from(other),
-    ///     Epoch::from_gregorian_utc(2022, 12, 01, 20, 21, 22, 0)
-    /// );
-    /// ```
-    pub fn with_hms_strict_from(&self, other: Self) -> Self {
-        let (sign, days, _, _, _, _, _, _) = self.duration.decompose();
-        let other = other.to_time_scale(self.time_scale);
-        Self::from_duration(
-            Duration::compose(
-                sign,
-                days,
-                other.hours(),
-                other.minutes(),
-                other.seconds(),
-                0,
-                0,
-                0,
-            ),
-            self.time_scale,
-        )
-    }
-
     pub fn month_name(&self) -> MonthName {
         let month = Self::compute_gregorian(self.duration, self.time_scale).1;
         month.into()
-    }
-
-    #[cfg(feature = "std")]
-    #[must_use]
-    /// Converts the Epoch to Gregorian in the provided time scale and in the ISO8601 format with the time scale appended to the string
-    pub fn to_gregorian_str(&self, time_scale: TimeScale) -> String {
-        let (y, mm, dd, hh, min, s, nanos) =
-            Self::compute_gregorian(self.duration, self.time_scale);
-
-        if nanos == 0 {
-            format!(
-                "{:04}-{:02}-{:02}T{:02}:{:02}:{:02} {}",
-                y, mm, dd, hh, min, s, time_scale
-            )
-        } else {
-            format!(
-                "{:04}-{:02}-{:02}T{:02}:{:02}:{:02}.{:09} {}",
-                y, mm, dd, hh, min, s, nanos, time_scale
-            )
-        }
     }
 
     #[cfg(feature = "std")]
@@ -2371,48 +1251,6 @@ impl Epoch {
                 "{:04}-{:02}-{:02}T{:02}:{:02}:{:02}.{:09}+00:00",
                 y, mm, dd, hh, min, s, nanos
             )
-        }
-    }
-
-    /// Returns the minimum of the two epochs.
-    ///
-    /// ```
-    /// use hifitime::Epoch;
-    ///
-    /// let e0 = Epoch::from_gregorian_utc_at_midnight(2022, 10, 20);
-    /// let e1 = Epoch::from_gregorian_utc_at_midnight(2022, 10, 21);
-    ///
-    /// assert_eq!(e0, e1.min(e0));
-    /// assert_eq!(e0, e0.min(e1));
-    /// ```
-    ///
-    /// _Note:_ this uses a pointer to `self` which will be copied immediately because Python requires a pointer.
-    pub fn min(&self, other: Self) -> Self {
-        if *self < other {
-            *self
-        } else {
-            other
-        }
-    }
-
-    /// Returns the maximum of the two epochs.
-    ///
-    /// ```
-    /// use hifitime::Epoch;
-    ///
-    /// let e0 = Epoch::from_gregorian_utc_at_midnight(2022, 10, 20);
-    /// let e1 = Epoch::from_gregorian_utc_at_midnight(2022, 10, 21);
-    ///
-    /// assert_eq!(e1, e1.max(e0));
-    /// assert_eq!(e1, e0.max(e1));
-    /// ```
-    ///
-    /// _Note:_ this uses a pointer to `self` which will be copied immediately because Python requires a pointer.
-    pub fn max(&self, other: Self) -> Self {
-        if *self > other {
-            *self
-        } else {
-            other
         }
     }
 }
@@ -2499,197 +1337,6 @@ impl FromStr for Epoch {
     }
 }
 
-impl fmt::Debug for Epoch {
-    /// Print this epoch in Gregorian in the time scale used at initialization
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let (y, mm, dd, hh, min, s, nanos) =
-            Self::compute_gregorian(self.duration, self.time_scale);
-        if nanos == 0 {
-            write!(
-                f,
-                "{:04}-{:02}-{:02}T{:02}:{:02}:{:02} {}",
-                y, mm, dd, hh, min, s, self.time_scale
-            )
-        } else {
-            write!(
-                f,
-                "{:04}-{:02}-{:02}T{:02}:{:02}:{:02}.{:09} {}",
-                y, mm, dd, hh, min, s, nanos, self.time_scale
-            )
-        }
-    }
-}
-
-impl fmt::Display for Epoch {
-    /// The default format of an epoch is in UTC
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let ts = TimeScale::UTC;
-        let (y, mm, dd, hh, min, s, nanos) =
-            Self::compute_gregorian(self.to_duration_in_time_scale(ts), ts);
-        if nanos == 0 {
-            write!(
-                f,
-                "{:04}-{:02}-{:02}T{:02}:{:02}:{:02} {}",
-                y, mm, dd, hh, min, s, ts
-            )
-        } else {
-            write!(
-                f,
-                "{:04}-{:02}-{:02}T{:02}:{:02}:{:02}.{:09} {}",
-                y, mm, dd, hh, min, s, nanos, ts
-            )
-        }
-    }
-}
-
-impl fmt::LowerHex for Epoch {
-    /// Prints the Epoch in TAI
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let ts = TimeScale::TAI;
-        let (y, mm, dd, hh, min, s, nanos) =
-            Self::compute_gregorian(self.to_duration_in_time_scale(ts), ts);
-        if nanos == 0 {
-            write!(
-                f,
-                "{:04}-{:02}-{:02}T{:02}:{:02}:{:02} {}",
-                y, mm, dd, hh, min, s, ts
-            )
-        } else {
-            write!(
-                f,
-                "{:04}-{:02}-{:02}T{:02}:{:02}:{:02}.{:09} {}",
-                y, mm, dd, hh, min, s, nanos, ts
-            )
-        }
-    }
-}
-
-impl fmt::UpperHex for Epoch {
-    /// Prints the Epoch in TT
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let ts = TimeScale::TT;
-        let (y, mm, dd, hh, min, s, nanos) =
-            Self::compute_gregorian(self.to_duration_in_time_scale(ts), ts);
-        if nanos == 0 {
-            write!(
-                f,
-                "{:04}-{:02}-{:02}T{:02}:{:02}:{:02} {}",
-                y, mm, dd, hh, min, s, ts
-            )
-        } else {
-            write!(
-                f,
-                "{:04}-{:02}-{:02}T{:02}:{:02}:{:02}.{:09} {}",
-                y, mm, dd, hh, min, s, nanos, ts
-            )
-        }
-    }
-}
-
-impl fmt::LowerExp for Epoch {
-    /// Prints the Epoch in TDB
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let ts = TimeScale::TDB;
-        let (y, mm, dd, hh, min, s, nanos) =
-            Self::compute_gregorian(self.to_duration_in_time_scale(ts), ts);
-        if nanos == 0 {
-            write!(
-                f,
-                "{:04}-{:02}-{:02}T{:02}:{:02}:{:02} {}",
-                y, mm, dd, hh, min, s, ts
-            )
-        } else {
-            write!(
-                f,
-                "{:04}-{:02}-{:02}T{:02}:{:02}:{:02}.{:09} {}",
-                y, mm, dd, hh, min, s, nanos, ts
-            )
-        }
-    }
-}
-
-impl fmt::UpperExp for Epoch {
-    /// Prints the Epoch in ET
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let ts = TimeScale::ET;
-        let (y, mm, dd, hh, min, s, nanos) =
-            Self::compute_gregorian(self.to_duration_in_time_scale(ts), ts);
-        if nanos == 0 {
-            write!(
-                f,
-                "{:04}-{:02}-{:02}T{:02}:{:02}:{:02} {}",
-                y, mm, dd, hh, min, s, ts
-            )
-        } else {
-            write!(
-                f,
-                "{:04}-{:02}-{:02}T{:02}:{:02}:{:02}.{:09} {}",
-                y, mm, dd, hh, min, s, nanos, ts
-            )
-        }
-    }
-}
-
-impl fmt::Pointer for Epoch {
-    /// Prints the Epoch in UNIX
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.to_unix_seconds())
-    }
-}
-
-impl fmt::Octal for Epoch {
-    /// Prints the Epoch in GPS
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.to_gpst_nanoseconds().unwrap())
-    }
-}
-
-#[must_use]
-/// Returns true if the provided Gregorian date is valid. Leap second days may have 60 seconds.
-pub const fn is_gregorian_valid(
-    year: i32,
-    month: u8,
-    day: u8,
-    hour: u8,
-    minute: u8,
-    second: u8,
-    nanos: u32,
-) -> bool {
-    let max_seconds = if (month == 12 || month == 6)
-        && day == usual_days_per_month(month - 1)
-        && hour == 23
-        && minute == 59
-        && ((month == 6 && july_years(year)) || (month == 12 && january_years(year + 1)))
-    {
-        60
-    } else {
-        59
-    };
-    // General incorrect date times
-    if month == 0
-        || month > 12
-        || day == 0
-        || day > 31
-        || hour > 24
-        || minute > 59
-        || second > max_seconds
-        || nanos > NANOSECONDS_PER_SECOND_U32
-    {
-        return false;
-    }
-    if day > usual_days_per_month(month - 1) && (month != 2 || !is_leap_year(year)) {
-        // Not in February or not a leap year
-        return false;
-    }
-    true
-}
-
-/// `is_leap_year` returns whether the provided year is a leap year or not.
-/// Tests for this function are part of the Datetime tests.
-const fn is_leap_year(year: i32) -> bool {
-    (year % 4 == 0 && year % 100 != 0) || year % 400 == 0
-}
-
 fn div_rem_f64(me: f64, rhs: f64) -> (i32, f64) {
     ((div_euclid_f64(me, rhs) as i32), rem_euclid_f64(me, rhs))
 }
@@ -2711,79 +1358,54 @@ fn rem_euclid_f64(lhs: f64, rhs: f64) -> f64 {
     }
 }
 
-#[test]
-fn div_rem_f64_test() {
-    assert_eq!(div_rem_f64(24.0, 6.0), (4, 0.0));
-    assert_eq!(div_rem_f64(25.0, 6.0), (4, 1.0));
-    assert_eq!(div_rem_f64(6.0, 6.0), (1, 0.0));
-    assert_eq!(div_rem_f64(5.0, 6.0), (0, 5.0));
-    assert_eq!(div_rem_f64(3540.0, 3600.0), (0, 3540.0));
-    assert_eq!(div_rem_f64(3540.0, 60.0), (59, 0.0));
-    assert_eq!(div_rem_f64(24.0, -6.0), (-4, 0.0));
-    assert_eq!(div_rem_f64(-24.0, 6.0), (-4, 0.0));
-    assert_eq!(div_rem_f64(-24.0, -6.0), (4, 0.0));
-}
+#[cfg(test)]
+mod ut_epoch {
 
-#[test]
-fn test_days_et_j2000() {
-    /*
-    WARNING: THIS ASSUMES THE UTC EPOCH in SPICE!
-    Verification via SPICE: load naif0012.txt (contains leap seconds until 2017-JAN-1)
-    In [6]: sp.str2et("2022-11-30 12:00:00")
-    Out[6]: 723081669.183061
-    In [7]: from hifitime import *
-    In [8]: Unit.Second*723081669.183061
-    Out[8]: 8369 days 1 min 9 s 183 ms 60 μs 992 ns @ 0x7fcd1559ef80
-    In [9]: (Unit.Second*723081669.183061).to_unit(Unit.Day)
-    Out[9]: 8369.000800729873
-    In [10]: (Unit.Second*723081669.183061).to_unit(Unit.Century)
-    Out[10]: 0.2291307542978747
+    use super::{div_rem_f64, Duration, Epoch};
 
-     */
-    let e = Epoch::from_tai_duration(Duration::from_parts(1, 723038437000000000));
-    let days_d = e.to_et_days_since_j2000();
-    let centuries_t = e.to_et_centuries_since_j2000();
-    assert!((days_d - 8369.000800729873).abs() < f64::EPSILON);
-    assert!((centuries_t - 0.2291307542978747).abs() < f64::EPSILON);
-}
-
-#[test]
-fn leap_year() {
-    assert!(!is_leap_year(2019));
-    assert!(!is_leap_year(2001));
-    assert!(!is_leap_year(1000));
-    // List of leap years from https://kalender-365.de/leap-years.php .
-    let leap_years: [i32; 146] = [
-        1804, 1808, 1812, 1816, 1820, 1824, 1828, 1832, 1836, 1840, 1844, 1848, 1852, 1856, 1860,
-        1864, 1868, 1872, 1876, 1880, 1884, 1888, 1892, 1896, 1904, 1908, 1912, 1916, 1920, 1924,
-        1928, 1932, 1936, 1940, 1944, 1948, 1952, 1956, 1960, 1964, 1968, 1972, 1976, 1980, 1984,
-        1988, 1992, 1996, 2000, 2004, 2008, 2012, 2016, 2020, 2024, 2028, 2032, 2036, 2040, 2044,
-        2048, 2052, 2056, 2060, 2064, 2068, 2072, 2076, 2080, 2084, 2088, 2092, 2096, 2104, 2108,
-        2112, 2116, 2120, 2124, 2128, 2132, 2136, 2140, 2144, 2148, 2152, 2156, 2160, 2164, 2168,
-        2172, 2176, 2180, 2184, 2188, 2192, 2196, 2204, 2208, 2212, 2216, 2220, 2224, 2228, 2232,
-        2236, 2240, 2244, 2248, 2252, 2256, 2260, 2264, 2268, 2272, 2276, 2280, 2284, 2288, 2292,
-        2296, 2304, 2308, 2312, 2316, 2320, 2324, 2328, 2332, 2336, 2340, 2344, 2348, 2352, 2356,
-        2360, 2364, 2368, 2372, 2376, 2380, 2384, 2388, 2392, 2396, 2400,
-    ];
-    for year in leap_years.iter() {
-        assert!(is_leap_year(*year));
+    #[test]
+    fn div_rem_f64_test() {
+        assert_eq!(div_rem_f64(24.0, 6.0), (4, 0.0));
+        assert_eq!(div_rem_f64(25.0, 6.0), (4, 1.0));
+        assert_eq!(div_rem_f64(6.0, 6.0), (1, 0.0));
+        assert_eq!(div_rem_f64(5.0, 6.0), (0, 5.0));
+        assert_eq!(div_rem_f64(3540.0, 3600.0), (0, 3540.0));
+        assert_eq!(div_rem_f64(3540.0, 60.0), (59, 0.0));
+        assert_eq!(div_rem_f64(24.0, -6.0), (-4, 0.0));
+        assert_eq!(div_rem_f64(-24.0, 6.0), (-4, 0.0));
+        assert_eq!(div_rem_f64(-24.0, -6.0), (4, 0.0));
     }
-}
 
-#[test]
-fn cumulative_days_for_month() {
-    assert_eq!(
-        CUMULATIVE_DAYS_FOR_MONTH,
-        [0, 31, 59, 90, 120, 151, 181, 212, 243, 273, 304, 334]
-    )
-}
+    #[test]
+    fn test_days_et_j2000() {
+        /*
+        WARNING: THIS ASSUMES THE UTC EPOCH in SPICE!
+        Verification via SPICE: load naif0012.txt (contains leap seconds until 2017-JAN-1)
+        In [6]: sp.str2et("2022-11-30 12:00:00")
+        Out[6]: 723081669.183061
+        In [7]: from hifitime import *
+        In [8]: Unit.Second*723081669.183061
+        Out[8]: 8369 days 1 min 9 s 183 ms 60 μs 992 ns @ 0x7fcd1559ef80
+        In [9]: (Unit.Second*723081669.183061).to_unit(Unit.Day)
+        Out[9]: 8369.000800729873
+        In [10]: (Unit.Second*723081669.183061).to_unit(Unit.Century)
+        Out[10]: 0.2291307542978747
 
-#[test]
-#[cfg(feature = "serde")]
-fn test_serdes() {
-    let e = Epoch::from_gregorian_utc(2020, 01, 01, 0, 0, 0, 0);
-    let content = r#""2020-01-01T00:00:00 UTC""#;
-    assert_eq!(content, serde_json::to_string(&e).unwrap());
-    let parsed: Epoch = serde_json::from_str(content).unwrap();
-    assert_eq!(e, parsed);
+         */
+        let e = Epoch::from_tai_duration(Duration::from_parts(1, 723038437000000000));
+        let days_d = e.to_et_days_since_j2000();
+        let centuries_t = e.to_et_centuries_since_j2000();
+        assert!((days_d - 8369.000800729873).abs() < f64::EPSILON);
+        assert!((centuries_t - 0.2291307542978747).abs() < f64::EPSILON);
+    }
+
+    #[test]
+    #[cfg(feature = "serde")]
+    fn test_serdes() {
+        let e = Epoch::from_gregorian_utc(2020, 01, 01, 0, 0, 0, 0);
+        let content = r#""2020-01-01T00:00:00 UTC""#;
+        assert_eq!(content, serde_json::to_string(&e).unwrap());
+        let parsed: Epoch = serde_json::from_str(content).unwrap();
+        assert_eq!(e, parsed);
+    }
 }
